refactor(cli): move configuration-related classes and context to config.py
feat(config): add new configuration management module

chore: update __all__ variable in cli.py
